Patch 1053
・PACKETVER「20120404」を追加 (vc_make.bat bcc32_make.bat Makefile clif.* mob.c)
	-> パケット0x977「モンスターHPバー更新」を実装
・PACKETVER「20120618」を追加 (vc_make.bat bcc32_make.bat Makefile clif.c)
	-> 状態異常アイコン表示に0x8ffではなく0x983を利用するようにする
・アイテム情報の装着箇所情報を修正
 (common/mmo.h, char/sql/chardb_sql.c, char/sql/maildb_sql.c, char/txt/chardb_txt.c, char/txt/maildb_txt.c,
　char/txt/storagedb_txt.c, converter/char-converter.c, converter/inter-converter.c)
・TXT:AURIGA_JOURNALを17に引き上げ (common/journal.c.)
・SQL:インベントリテーブルの装着箇所情報を修正（sql-files/main.sql, sql-files/Auriga1053_changetable.sql）
・テーブル定義更新 (db/packet_db.txt clif.c skill.h status.h)
・client_packet.txtを少し追記と修正 (doc/client_packet.txt)

+++ src/char/sql/chardb_sql.c
@@ -85,7 +85,7 @@ int chardb_sql_loaditem(struct item *item, int max, int id, int tableswitch)
 			item[i].id        = (unsigned int)atoi(sql_row[0]);
 			item[i].nameid    = atoi(sql_row[1]);
 			item[i].amount    = atoi(sql_row[2]);
-			item[i].equip     = atoi(sql_row[3]);
+			item[i].equip     = (unsigned int)atoi(sql_row[3]);
 			item[i].identify  = atoi(sql_row[4]);
 			item[i].refine    = atoi(sql_row[5]);
 			item[i].attribute = atoi(sql_row[6]);
