+++ src/char/txt/chardb_txt.c
@@ -507,7 +507,7 @@ static int mmo_char_fromstr(char *str,struct mmo_chardata *p)
 			p->st.inventory[i].id        = (unsigned int)tmp_int[0];
 			p->st.inventory[i].nameid    = tmp_int[1];
 			p->st.inventory[i].amount    = tmp_int[2];
-			p->st.inventory[i].equip     = tmp_int[3];
+			p->st.inventory[i].equip     = (unsigned int)tmp_int[3];
 			p->st.inventory[i].identify  = tmp_int[4];
 			p->st.inventory[i].refine    = tmp_int[5];
 			p->st.inventory[i].attribute = tmp_int[6];
@@ -551,7 +551,7 @@ static int mmo_char_fromstr(char *str,struct mmo_chardata *p)
 			p->st.cart[i].id        = (unsigned int)tmp_int[0];
 			p->st.cart[i].nameid    = tmp_int[1];
 			p->st.cart[i].amount    = tmp_int[2];
-			p->st.cart[i].equip     = tmp_int[3];
+			p->st.cart[i].equip     = (unsigned int)tmp_int[3];
 			p->st.cart[i].identify  = tmp_int[4];
 			p->st.cart[i].refine    = tmp_int[5];
 			p->st.cart[i].attribute = tmp_int[6];
